BACKPORT: drm/i915: set FORCE_ARB_IDLE_PLANES workaround
Commit 1544d9d57396d5c0c6b7644ed5ae1f4d6caad07a added a workaround
inside haswell_init_clock_gating and mentioned it is "a workaround for
early silicon revisions and should be removed later". This workaround
is documented in bit 31 of PRI_CTL. I asked Arthur and he mentioned
that setting FORCE_ARB_IDLE_PLANES replaces that workaround for the
newer machines. So use the new one.

Also notice that there's still another workaround for PRI_CTL that
involves WM_DBG, but it's not the one we're reverting. And notice that
we were previously setting WM_DBG_DISALLOW_MULTIPIPE_LP which disables
the LP watermarks when more than one pipe is used, and we really don't
want this because we need the LP watermarks if we want to reach deeper
PC states.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
(cherry picked from commit 6ee64ed0b8398982c43699de089410cd75d794f0)

Conflicts:
	drivers/gpu/drm/i915/intel_pm.c
(context change)

BUG=chrome-os-partner:19682
TEST=With entire patch series applied, observe PC7 residency via powertop in
     crosh.
Change-Id: I4212e9910738f8876c088c0fe7e8f5736e1ba94a
Reviewed-on: https://gerrit.chromium.org/gerrit/56485
Tested-by: Josh Triplett <josh@joshtriplett.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>