CHROMIUM: config: Remove dummy regulator for exynos
Now that we have tpschrome in place we can remove the dummy regulator
hack that we had in.  This allows us to properly probe for regulators
that will be available later.

NOTE: I think this may break HDMI since it doesn't specify regulators
properly and we don't have our old patch to the HDMI driver to have it
avoid grabbing its regulators.

BUG=chromium:239532
TEST=Boot and can see LCD come up.

Change-Id: I60d1c3ecec221023782f10d98c21a2c7dbc26c11
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/49904