google memconsole: try direct pointer in ACPI
Provide a way to obtain the coreboot memconsole pointer directly
from an ACPI field named 'CBMC'. Boards using the dynamic cbmem
in coreboot have a different structure for the layout coreboot
memory. Instead of writing another parser for the new layout rely
on the presence of a 'CBMC' field in the ACPI tables to provide
a pointer to the memory console.

BUG=None
BRANCH=None
TEST=Manually verfied /sys/firmware/log being populated with the
     coreboot memory console. Also verified that cbmem from upstream
     works correctly as well.

Change-Id: I53e659d5733837f885baf4c5a087a2c44c92d437
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/48851