CHROMIUM: Input: cyapa - Retry power mode reads and writes
There are some conditions where the trackpad is unresponsive to a power
mode transition, especially when the trackpad is in Off state.
Allow for 5 retries with a 10ms delay.

This change is refactored from chromeos-3.4, and rolls together the
following two commits :
78ea988 CHROMIUM: Input: cyapa - retry power mode writes
4c9a20e CHROMIUM: Input: cyapa - Retry power mode reads

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:8070,chrome-os-partner:9229,chromium:221149
TEST=Suspend and resume the system several times with the lid open
(using powerd_suspend) or with the lid closed, by closing the lid.
Ensure that the following errors do not appear in the logs:
set power mode failed, -6
resume active power failed, -6
set active power failed, -6

Change-Id: I8feac5adaa0f394cd19fe338a223e147c6269819
Reviewed-on: https://gerrit.chromium.org/gerrit/49379
Reviewed-by: Tai-Hsu Lin <sheckylin@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>