drm/i915: put context upon switching
In order to be notified of when the context and all of its associated
objects is idle (for if the context maps to a ppgtt) we need a callback
from the retire handler. We can arrange this by using the kref_get/put
of the context for request tracking and by inserting a request to
demarque the switch away from the old context.

BUG=chrome-os-partner:22147
TEST=Sanity Test on Link(unable to repro)

[Ben: fixed minor error to patch compile, AND s/last_context/from/]
Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>

(cherry picked from commit 112522f6789581824903f6f72082b5b841a7f0f9)
Signed-off-by: David Sodman <dsodman@chromium.org>

Change-Id: Ia0367621844f597136802c0b54e40e16064abc48
Reviewed-on: https://chromium-review.googlesource.com/181029
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: David Sodman <dsodman@chromium.org>