ARM: dts: Add i2c-arbitrator for exynos5250-snow and move devices
We need to use the i2c-arbitrator to talk to any of the devices on i2c
bus 4 on exynos5250-snow so that we don't confuse the embedded
controller (EC).  Add the i2c-arbitrator to the device tree.

This is not a very "pure" CL in that it takes out the device tree bits
related to the old-style arbitration and adds in the new-style
arbitration.  It also moves devices around since we already had a
bunch of stuff listed under bus 4.  Upstream is getting devices
cleanly put under the arbitration bus to begin with.

The arbitrated bus is numbered 104 simply as a convenience to make it
easier for people poking around to guess that it might have something
to do with the physical bus 4.

BUG=chromium:239532
TEST=Boot and see cros_ec and tpschrome under bus 104.

Change-Id: I2ba28ec2dcb6d261c4651042a8167b4360ca7595
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/49907
Reviewed-by: Olof Johansson <olofj@chromium.org>