BACKPORT: drm/i915: fix panel fitting on LVDS on ILK+ v2
This regression was introduced in:

commit b074cec8c652f2d273907a4b35239b4766c894ac
Author: Jesse Barnes <jbarnes@virtuousgeek.org>
Date:   Thu Apr 25 12:55:02 2013 -0700

    drm/i915: move PCH pfit controls into pipe_config

In refactoring this, it was only applied to eDP, which is incorrect.  In
fact, if we ever use the panel fitter to deal with overscan on HDMI,
we'll need to extend it again, so just drop the conditional altogether.

v2: drop check for eDP since we can use the fitter in any config (Daniel)

Signed-off-by: Jesse Barnes <jbarnes@virtuousgeek.org>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 0ef37f3f5e33eae7d6c388a7b374397794beca39)
Signed-off-by: Yuly Novikov <ynovikov@chromium.org>

BUG=chrome-os-partner:19682
TEST=compiles

Change-Id: Icea0bf293c56343a1564dbc5c0fe74c655ea4210
Reviewed-on: https://gerrit.chromium.org/gerrit/59952
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Tested-by: Yuly Novikov <ynovikov@chromium.org>