CHROMIUM: kcrash dump to core file converter
kcrash-to-core converts information in the log file from a crash dump
(typically in /dev/pstore/console-ramoops) to a core file which may
be used in conjunction with the debug kernel to extract line number
information, as well as the values of some local variables.

With the -f flag, the program tries to fix stack values that look like
code addresses to compensate for the offset introduced by kaslr.

BUG=chromium:238491
TEST=produced a core file from console-ramoops and ran gdb <vmlinux> <core>

Change-Id: I357e5651b0d4dc1ade8479df46460d88edcc8b41
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/50238
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>