drm/exynos: use SoC name to identify hdmi version
Exynos hdmi IP version is named after hdmi specification version i.e.
1.3 and 1.4. This versioning mechanism is not sufficient to handle
the diversity in the hdmi/phy IPs which are present across the exynos
SoC family.

This patch changes the hdmi version to the name of the SoC in which
the IP was introduced for the first time. Same version is applicable
to all subsequent SoCs having the same IP version.

BUG=chrome-os-partner:19044
TEST=basic hdmi works fine on Snow, smdk5420, peach board.

Signed-off-by: Rahul Sharma <rahul.sharma@samsung.com>

Change-Id: I75f250d232ad528dea488f83f0ceff48e9d1ac3a
Reviewed-on: https://gerrit.chromium.org/gerrit/56987
Commit-Queue: Shirish S <shirish@chromium.org>
Reviewed-by: Shirish S <shirish@chromium.org>
Tested-by: Shirish S <shirish@chromium.org>