CHROMIUM: verity: fix UUID checking for VMTests
Two issues came up when running 3.8 on a VM.  First was that the
arguments to verity were slightly different then when running on
ChromeOS hardware, because the UUIDs had "PARTUUID=" pre-pended to
them, which is correct form for the Linux kernel UUID code.
This form of UUID actually takes a slightly different parsing path
through dm-table, which hadn't been updated as part of the rebase.
Second, as part of the 3.8 rebase, I had to change the code which found
partitions by UUID because the kernel switched to storing them in
ASCII, and the original change made the UUIDs case-sensitive which
isn't correct behavior.  Finally, clean up some memory allocation
issues with the original uuid matching changes.

Hopefully a lot of this duplication and churn will get cleaned up when
we refactor this as part of crbug.com/224066.

BUG=chromium:238326
TEST=boots on vm
TEST=boots on ChromeOS hardware
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>

Change-Id: I3e10fef8deb6f20f78311980f20ff6164f79efb4
Reviewed-on: https://gerrit.chromium.org/gerrit/50402
Reviewed-by: Paul Taysom <taysom@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>