TEMP: ARM: exynos: Select PINCTRL_EXYNOS for exynos5
If you've got MACH_EXYNOS5_DT but not MACH_EXYNOS4_DT you'll be
missing the pincontrol definitions.  Add them for exynos5.

This is a backport of what's been posted to LKML:
  https://patchwork.kernel.org/patch/2540901/
It's a minor backport since our Kconfig looks a bit different from
upstream.  I've alphabetized the list at the same time.

BUG=chrome-os-partner:19182
TEST=build and booted on snow till login screen

Change-Id: I5688e9c5f578a4b3633a7109ef8babb6fa9e240b
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/50414
Reviewed-by: Olof Johansson <olofj@chromium.org>