BACKPORT: mwifiex: fix setting of multicast filter
A small bug in this code was causing the ALLMULTI filter to be set
when in fact we were just wanting to program a selective multicast list
to the hardware.

Fix that bug and remove a redundant if condition in the code that
follows.

This fixes wakeup behaviour when multicast WOL is enabled. Previously,
all multicast packets would wake up the system. Now, only those that the
host intended to receive trigger wakeups.

Signed-off-by: Daniel Drake <dsd@laptop.org>
Cc: <stable@vger.kernel.org>
Acked-by: Bing Zhao <bzhao@marvell.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>

BUG=chrome-os-partner:37047
TEST=The issue has been identified and root caused as device
sends deauth during traffic in response to AP's multicast deauth
frame. With our hacked AP which sends multicast deauth whenever
needed, we could recreate the scenario and verify this fix.

Change-Id: I5067adee6e83f5e9f32d2777d0bc2dddf4273972
Signed-off-by: Amitkumar Karwar <akarwar@marvell.com>
Reviewed-on: https://chromium-review.googlesource.com/266831
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Avinash Patil <patila@marvell.com>
Tested-by: Avinash Patil <patila@marvell.com>