drm/exynos: Remove dpms link between encoder/connector
This patch removes the call from encoder dpms into connector dpms (which
will then call back into encoder dpms through the helper function). The
callback is likely to keep connector->dpms in the right state when
initiating dpms from crtc or encoder, but this isn't the right way to do
it. This patch is the first step towards rationalizing power management
in the exynos drm driver.

This patch also removes the apply() callback from the dpms path. It's
incorrect to assume apply should be called from dpms. If it's a
requirement for the individual hardware drivers, they should call it
internally.

BUG=chromium:234847
TEST=Tested on snow, no regressions noted

Change-Id: I3805dfc94b739075a793576a06eee0873ada8773
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/56284