CHROMIUM: drm/exynos: hdmi: don't readback hdmiphy_conf
These registers are only read for debugging, it is a waste of time to read
them back if we aren't going to print them...

But the real kicker is, since the hdmiphy is a write-only i2c interface,
the registers always read as 0.  I'm not sure how this was never noticed
by whomever originally implemented this.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=none

Change-Id: I59f1e6f9c74ae609f65663ff0d447986b90f734e
Reviewed-on: https://gerrit.chromium.org/gerrit/50602
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>