CHROMIUM: drm/bridge/ps8620: Support for multiple DP lanes to PS8622
The PS8625 bridge chip has support for running with 1 or 2 lanes active. This
sets the maximum number of lanes on the bridge chip based on the compat string
from the device tree. It also adds a property for directly setting the number of
lanes in the device tree node for the bridge chip because there is no general
way to query the display port controller to see how many lanes it is using.

Original-Change-Id: Ibd817f7911023d6b3f7f0ff2745375647e57cf8b
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/37667
Reviewed-by: Simon Glass <sjg@google.com>
(cherry picked from peach-alpha commit
	39a1d6bd97ba1fa524e72e0754eba5b9a50c12f0)

BUG=chrome-os-partner:19760
TEST=Build and boot on snow

Change-Id: Ic4f9030eeb947888bbf5a80992997cd8c8c7e4bd
[abrestic: fixed up comments]
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/55928