clk/exynos5250: add mout_hdmi mux clock for hdmi
hdmi driver needs to change the parent of hdmi clock
frequently between pixel clock and hdmiphy clock. hdmiphy is
not stable after power on and for sometime after changing the
phy configuration. For this duration pixel clock is used to
clock hdmi.

This patch is exposing the mux for changing parent.

BUG=chrome-os-partner:19785
TEST=basic hdmi works fine on Snow, smdk5420, peach board.

Change-Id: I6c0481d0a191b99e768f6c40b25b5fe33d0e7359
Signed-off-by: Rahul Sharma <rahul.sharma@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/56974
Commit-Queue: Doug Anderson <dianders@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>