CHROMIUM: usb: samsung-usb2phy: Toggle HSIC GPIO from device tree
This patch adds support for a new 'samsung,hsic-reset-gpio' in the
device tree, which will be interpreted as an active-low reset pin during
PHY initialization when it exists. Useful for intergrated HSIC devices
like the SMSC3503 hub.

Also fixes PHY semaphore code to make sure we always go through the
setup at least once, even if it was already turned on (e.g. by
firmware).

BUG=chrome-os-partner:20205
TEST=Run 'lsusb' on Pit, make sure 0424:3503 shows up. Do an S2R cycle
and make sure it's still there.

Original-Change-Id: I2a8cee0e036a8fdf605129447a9b15191a64e22d
Change-Id: I402f12901bdcc8bca41786fc6a4c348743fc61b1
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58963
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>