drm/exynos: replace dummy hdmiphy clock with pmu register control
Previous to CCF, hdmiphy is added as a dummy clock in clock file for
exynos SoCs. Enable/Disable to this clock, actually toggles the power
control bit in PMU, instead of controlling the clock gate.

Patch adds the support to parse hdmiphy control node which is a child
node to hdmi, and map the pmu register to toggle the power control bit.

BUG=chrome-os-partner:19785
TEST=basic hdmi works fine on Snow, smdk5420, peach board.

Signed-off-by: Rahul Sharma <rahul.sharma@samsung.com>

Change-Id: I96267cef34b80920542d1f14d837bcd8b85f295b
Reviewed-on: https://gerrit.chromium.org/gerrit/56977
Commit-Queue: Shirish S <shirish@chromium.org>
Reviewed-by: Shirish S <shirish@chromium.org>
Tested-by: Shirish S <shirish@chromium.org>