CHROMIUM: drm/exynos: remove vsync wait from destroy_fb
The vsync wait is now unneccesary since we now have proper
refcounting in place.

BUG=chrome-os-partner:15349,chrome-os-partner:14965
TEST=Multiple VT switch, HDMI plug/unplug, sign in/out, suspend/resume,
     idle suspend/resume.

Original-Change-Id: Idb5fe8615021d35db3e5a4c6a528e8386a4d84d7
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/37575

[djkurtz: 3.8 rebase: Many conflicts, but the core idea remains:
no need to wait for vblank (in 3.8 'complete_scanout') in the fb release
worker, b/c we only unref the fb when we already know that it isn't being
scanned out - only from error paths before it is applied, or from the
vblank irq after it has just been page flipped out.  And, since this was
the only use of 'complete_scanout', remove it and all its ops callbacks]

Change-Id: Ib817faf087ee8d268c1642ced6cc2e459a6e15ae
Reviewed-on: https://gerrit.chromium.org/gerrit/50435
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>