ARM: dts: Configure pulls and drive strengths for exynos5250-snow
When moving from the old Samsung GPIO specifier to the new glorious
pinmux world, we no longer specify the pullup/pulldown nor the drive
strength of lines directly where we refer to them in the device tree.
Instead we need to use pinctrl to do it.

We'll take advantage of the fact that the pinctrl code will
automatically setup the default state of pins when a device probes and
associate pins with the devices.

BUG=chrome-os-partner:19317
TEST=Can now use power button, lid switch, and run
suspend_stress_test -c1 --suspend_min=20
TEST=Compared pull and drive strength values to what they were in
chromeos-3.4 for all affected values and verified that they are the
same or better.

Change-Id: Ie151ad54141c217d3a52edfa6f4545efbffc9d73
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/51232
Reviewed-by: Olof Johansson <olofj@chromium.org>