Revert "pinctrl: samsung: fix suspend/resume functionality"
This reverts commit 804a790cc1d56a7dadb33a840377ed56bf0b820f.

There was nothing wrong with the above commit, but it was based on an
older version of the upstream pinmux code upstream which makes it a
pain to send upstream.  Since all of the rest of the suspend/resume
code hasn't been sent yet, let's:
* Revert it
* Apply the newer upstream patches
* Submit a newer version that matches upstream changes.
  See Ife08f1a377ce14fbb1550e80f27f0dc503b56e93

BUG=chrome-os-partner:19317
TEST=Still compiles with this change reverted; with all changes
suspend/resume is working much better.

Change-Id: I19250063f544b1e8bfddad9ecf32f5fcfb76a637
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/51328
Reviewed-by: Prathyush Kalashwaram <prathyush@chromium.org>
Tested-by: Prathyush Kalashwaram <prathyush@chromium.org>