Revert "pinctrl: exynos: add irq_set_wake function"
This reverts commit 05258d3ed6a32cb8611dc04b2d7084983a6b49f5.

There was nothing wrong with the above commit, but it was based on an
older version of the upstream pinmux code upstream which makes it a
pain to send upstream.  Since all of the rest of the suspend/resume
code hasn't been sent yet, let's:
* Revert it
* Apply the newer upstream patches
* Submit a newer version that matches upstream changes.
  See I85a67b237d6c5f46760ee5a77023502b218a818f

BUG=chrome-os-partner:19317
TEST=Still compiles with this change reverted; with all changes
suspend/resume is working much better.

Change-Id: I2c27d806a12f8ab499940abec70dff1dd1afaf13
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/51341
Reviewed-by: Prathyush Kalashwaram <prathyush@chromium.org>
Tested-by: Prathyush Kalashwaram <prathyush@chromium.org>