TEMP: ARM: dts: Add powerdown config to exynos5250-snow dts file
We take advantage of the new ability added in '71ea086 pinctrl:
samsung: add ability to set powerdown configs' to specify the
powerdown config for snow.  This is important for proper functioning
of suspend/resume.

This change is only temporary, as we should really be using the bindings
already present upstream ("samsung,pin-{pud,con}-pdn") to specify the
powerdown settings.

BUG=chrome-os-partner:19317
BUG=chrome-os-partner:19234
TEST=Suspend system using powerd_suspend and wake up system using key
press and display comes back normally.
TEST=Dump all powerdown settings and spot check; double check that
values in device tree match what we had for 3.4.

Change-Id: I71895ac359e9601ebe13b19c47dfa20041e108a9
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Signed-off-by: Jonathan Kliegman <kliegs@chromium.org>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/51372