UPSTREAM: clk: exynos5250: Add sclk_mpll to the parent list of mout_cpu clock
'mout_mpll' is added the list of parent clocks for 'mout_cpu'.
'mout_mpll' is an alias to the clock 'sclk_mpll'. Hence 'sclk_mpll'
should be added to the list of parent clocks.

This results in an error when cpufreq driver for EXYNOS5250 tries to
set 'mout_mpll' as a parent for 'mout_cpu'.

clk_set_parent: clk sclk_mpll can not be parent of clk mout_cpu

Signed-off-by: Tushar Behera <tushar.behera@linaro.org>
Signed-off-by: Mike Turquette <mturquette@linaro.org>
Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:21952
TEST=boot on Spring, check in the kernel log that we no longer have
exynos_cpufreq initialization errors and see the CPU frequency going up
and down.

Change-Id: I803f99160948162c27f4b6954c4dfac89ab8dfb4
Reviewed-on: https://gerrit.chromium.org/gerrit/66225
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>