UPSTREAM: ASoC: max98090: request IRQF_ONESHOT interrupt
request_threaded_irq() rejects calls which both do not specify a handler
(indicating that the primary IRQ handler should be used) and do not set
IRQF_ONESHOT because the combination is unsafe with level-triggered
interrupts.  It is safe in this case, though, since max98090 IRQs are
edge-triggered and the interrupts aren't ACK'ed until the codec's IRQ
status register is read.  Because of this, an IRQF_ONESHOT interrupt
doesn't really make a difference, but request one anyway in order to make
request_threaded_irq() happy.

BUG=chrome-os-partner:18720
TEST=Once config option and DT bindings are added; max98090 probes
successfully.

Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>
(cherry picked from linux-next commit
	3d15aacbb802af72b4ff0c3ba576536cdb3bace0)

Change-Id: I0a0df46863c11ce3b8e44f669381d70c23d1e74b
Reviewed-on: https://gerrit.chromium.org/gerrit/57631
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>