cpufreq: exynos: allow to set default maximum frequency from device tree
When the cpu node in the device tree has the "clock-frequency-limit" attribute
set, the cpufreq policy maximum frequency is set to the given frequency
value.
This allows the firmware to give a hint about what is the acceptable
operating range to avoid exceeding the available power envelope.
The setting can be reset from userspace by using the cpufreq sysfs
interface.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:18927
TEST=on Spring, run with a firmware passing clock-frequency-limit=800000 in
the device tree, and see the "Capping CPU0 frequency to 800 Mhz" trace
in the kernel log and the CPU frequency not going higher than 800Mhz
after startup.

Change-Id: I89a40b977ad56c05ce53d47baa8d78da0cf3c5c3
Reviewed-on: https://gerrit.chromium.org/gerrit/55790
Commit-Queue: Todd Broch <tbroch@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>