CHROMIUM: clk: exynos5250: Add an alias for CHIP ID so ASV can find it
The ASV driver isn't device tree enabled and doesn't appear to have a
dev node so there's no way to give it a reference to its clock without
making it global.  For now we'll do that and we'll fix soon.

See http://crosbug.com/p/19646 for task to remove this.

BUG=chrome-os-partner:19183
TEST=With next patch too, boot and see proper chip ID / lot ID in
dmesg (compare to 3.4).  On my machine:
  [    1.100000] EXYNOS5250: Adaptive Support Voltage init
  [    1.170000] Exynos5250: Lot ID is NZW25 and wafer number is 2
  [    1.175000] EXYNOS5250: IDS:21 HPM:15 RESULT:7 MIF:2

Change-Id: I14c6ade1d02053cc57e1d0ded994548f965948ea
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/56129